fix: use Identity and Relationship attribute renderers for response renderers
