+++ packages/renderers/lib/src/renderers/request_item_renderer/response/share_attribute_accept_response_item_renderer.dart
@@ -1,7 +1,8 @@
 import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/material.dart';
 
-import '../../widgets/custom_list_tile.dart';
+import '../widgets/identity_attribute_value_renderer.dart';
+import '../widgets/relationship_attribute_value_renderer.dart';
 
 class ShareAttributeAcceptResponseItemRenderer extends StatelessWidget {
   final ShareAttributeAcceptResponseItemDVO item;
@@ -10,9 +11,10 @@ class ShareAttributeAcceptResponseItemRenderer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return CustomListTile(
-      title: 'i18n://dvo.attribute.name.${item.attribute.value.atType}',
-      description: item.attribute.value.toJson()['value'],
-    );
+    return switch (item.attribute.value) {
+      final IdentityAttributeValue value => IdentityAttributeValueRenderer(value: value),
+      final RelationshipAttributeValue value => RelationshipAttributeValueRenderer(value: value),
+      _ => throw Exception('Unknown AttributeValue: ${item.attribute.valueType}'),
+    };
   }
 }
