+++ packages/renderers/lib/src/renderers/request_item_renderer/widgets/relationship_attribute_value_renderer.dart
@@ -2,10 +2,10 @@ import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/widgets.dart';
 
 class RelationshipAttributeValueRenderer extends StatelessWidget {
-  final List<IdentityAttributeDVO> results;
+  final List<IdentityAttributeDVO>? results;
   final RelationshipAttributeValue value;
 
-  const RelationshipAttributeValueRenderer({super.key, required this.value, required this.results});
+  const RelationshipAttributeValueRenderer({super.key, required this.value, this.results});
 
   @override
   Widget build(BuildContext context) {
